# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Academic World Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /neo4j/collaboration-network:
    get:
      operationId: neo4jCollaborationNetwork
      tags: [neo4j]
      summary: Co-author collaboration network
      parameters:
        - in: query
          name: facultyName
          required: true
          schema:
            type: string
        - in: query
          name: depth
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
  /neo4j/faculty-keywords:
    get:
      operationId: neo4jFacultyKeywords
      tags: [neo4j]
      summary: Keywords associated with a faculty member
      parameters:
        - in: query
          name: facultyName
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyKeywordScore"
  /neo4j/keyword-experts:
    get:
      operationId: neo4jKeywordExperts
      tags: [neo4j]
      summary: Top faculty experts for a keyword
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyScore"
  /neo4j/research-path:
    get:
      operationId: neo4jResearchPath
      tags: [neo4j]
      summary: Shortest collaboration path between two faculty
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
        - in: query
          name: to
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchPath"
  /neo4j/network-centrality:
    get:
      operationId: neo4jNetworkCentrality
      tags: [neo4j]
      summary: Most connected faculty by degree centrality
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CentralityResult"
  /neo4j/research-clusters:
    get:
      operationId: neo4jResearchClusters
      tags: [neo4j]
      summary: Discover research clusters via graph connected components
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClusterMember"
components:
  schemas:
...
    FacultyScore:
      type: object
      properties:
        name:
          type: string
        university:
          type: string
        score:
          type: number
      required: [name, score]
    FacultyKeywordScore:
      type: object
      properties:
        keyword:
          type: string
        score:
          type: number
      required: [keyword, score]
...
    GraphData:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
      required: [nodes, edges]
    GraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        university:
          type: string
      required: [id, label]
    GraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        weight:
          type: integer
      required: [source, target]
    ResearchPath:
      type: object
      properties:
        path:
          type: array
          items:
            type: string
        length:
          type: integer
      required: [path, length]
    CentralityResult:
      type: object
      properties:
        name:
          type: string
        university:
          type: string
        degree:
          type: integer
        sharedPubs:
          type: integer
      required: [name, degree, sharedPubs]
    ClusterMember:
      type: object
      properties:
        name:
          type: string
        university:
          type: string
        score:
          type: number
      required: [name, score]
...